Zombie now moves(yayyyy) + Some Minor Changes

+++ Bullet.cpp
@@ -1,5 +1,5 @@
 #include "Bullet.h"
-Bullet::Bullet(int x, int y, float bulletSpeed, string spritePath, int textureX, int textureY): pos(x,y),bulletSpeed(bulletSpeed) {
+Bullet::Bullet(int x, int y, float bulletSpeed, string spritePath, double textureX, double textureY): pos(x,y),bulletSpeed(bulletSpeed) {
 	bulletTexture.loadFromFile(spritePath);
 	bulletSprite.setTexture(bulletTexture);
 	bulletSprite.setTextureRect(IntRect(0,0,textureX,textureY));
